fix: /mypage/_components 만들어서 컴포넌트 분리
